+++ NEWS.md
@@ -1,5 +1,12 @@
 # MsExperiment 0.99
 
+## MsExperiment 0.99.2
+
+- Fix imports.
+- Inherit from the `S4Vectors::Annotated` class.
+- Rename the `@assay` slot into `@qdata`.
+- Fix class unions.
+
 ## MsExperiment 0.99.0
 
 - Prepare for Bioconductor submission.